DYN-5656 Dynamo UI Blocking Action
I've implementing the functionality of adding the GuideBackground (overlay) when the Preferences panel or the PackageSearchWindow are opened and removing it when those windows are closed. In this way when we have the GuideBackground the user won't be able to interact with the background.
I've added this functionality in the DynamoView but due that the GuideBackground is not working with the Library, I had to implement the overlay functionality in the library.html and the functions needed to set the overlay from .NET.