DYN-5791 Fixing Renaming Tool
If the md file name contains empty spaces then the img file name will also contain %20 instead of empty space so in this way will be loaded by DocumentationBrowser correctly.
So my fix is to remove the %20 when renaming the files and the md file content so matches the expected hashed img name.