DYN-5656-PackagesGuide-BugFix
When opening the Packages tour and later opening the PackageManager (during the tour) at closing was removing the wrong dark background, then I added some code for searching for the specific overlay added  (not the one added by the guided tour).
Also deleted the code for using the EnableEnvironment function due that  is not used any more.