DYN-6603 Decrease Weight T-Spline Nodes
I've changed the weight of Spline nodes at indexing time so when using criteria like "sphere" will show first the nodes under the Sphere category and later the T-Spline nodes which has the text "sphere" in the name.
For implementing  the expected behavior I've decreased the weight of the fields Description and SearchKeywords just for t-spline nodes.