DYN-6995 Crash When Changing Language
This problem is happening only in Dynamo Revit and happens due that the DynamoCore.dll is not being localized when inserting a graph (which contains a group) is trying to find a default group with a localized name but is not found then is making the code to crash.
With this fix I'm forcing to assign a group even when the default group is not found.