DYN-6569 Extended Characters Md2HTML Fix
After debugging more and inserting logs inside the Md2Html.exe tool I notice that at the beginning when chosing the UTF-8 Encoding we were sending some weird characters so instead of using == now I'm using .Equals().
So I deleted the previous code for encoding and decoding the content and set to create the process with UTF-8.