DYN-7273 Search Issue Civil3D
When a user in Civil3D was searching for a node like "choose text style" no results were not found in the top ten nodes.

This problem was happening due that when the SearchTerm contains empty space the algorithm parse the term in tokens so matches all the nodes which contain "choose", "text" and "style" in the node name and due that there is a restriction that the if the number of matching nodes > 20 then discard that case so is not showing results for that searchTerm.

Also there was a missing condition that when the term has spaces we should try to match the name with spaces, this was the code added in this commit and this is why now we are getting results when searching for "choose text style".