DYN-7273 Search Issue Civil3D
I realized that with my fix searching for "set parameter" in Revit was moving the SetParameterByName node from the 1st position to the 5th position then for fixing that problem I've modified the code in a way that will be using the default weight for Wildcard Query (instead of using the once calculated by the CalculateFieldWeight() method),

Also was reported that when searching for "file path" was not showing the File Path node in the list, it was happening due that the word "file" is a Category so the results were showing the nodes under the File Category that contain the word "path" in the Name. For fixing this problem I removed the MUST for the category so it will show also extra nodes in the results (including the File Path node).