DYN-7365 Updating Lucene Search Algorithm
Updating the Weight for the Fields: Description, SearchKeywords due that when searching a term like "combine" there are several nodes that contain the word "combine" in the SearchKeywords and those nodes are above than the expected node (List.Combine).

Also I've added several test that validates the search for specific nodes.