DYN-7442 Search Nodes with Special Characters
When Searching nodes with special characters in the Name when Dynamo is in a different language than English was not getting the expected results, the problem was related to the Analyzer being used.
In this fix I'm using the CustomAnalyzer providing the Dynamo language as a parameter, this will allow using the expected character set.