DYN-7480 Merge master_cl-localization into master_cl (#15493)
After a several tests I realized that the .resx files with localized strings can be added directly to the project directory location (csproj) level without the need of using Visual Studio. Usually when using Visual Studio for adding resources the <EmbeddedResource> tag is added inside the .csproj file but in this case after testing the localization for other languages was working as expected without using Visual Studio for adding resources.
Also the Access Modifier for resx files localized can have the value "No code generation" and the localization in still working.