DYN-7647 Search PackageManager EmptySpace Regression
After debugging the tests I found that If I run just the test it passes but if I run all the TestFixture the PackageSearchWithWhitespaceInName() test was failing. The reason was that for LuceneSearch.LuceneUtilityPackageManager we were using only one instance for all the tests so in a previous test all the info was indexed and committed so when reaching the LuceneSearch.LuceneUtilityPackageManager() test the new indexed info was not reflected (even when doing the commit).
For fixing this problem I assigned LuceneSearch.LuceneUtilityPackageManager = null so this will create again the LuceneUtility instance and the indexed info will be reflected successfully.