DYN-7435 Corrupted Characters in DocBrowser
When using a Windows 10 OS installed in Chinese Traditional if you open DocumentationBrowser several times for each node in the workspace, the DocumentationBrowser starts to show weird characters and the html layout seems different, this was happening due that the Md2Html.Convert and Md2Html.Sanitize methods were not executed, after my fix now they were executed but was still showing weird characters (due that Md2Html encoding was in Chinese Traditional). Then for fixing this problem I had to encode the info using base64 and decoded when is needed.
There was a freeze reported in the Jira task but I was not able to reproduce it after my fix.