DYN-7435 Corrupted Characters in DocBrowser Regressions
Due that FeatureFlags is using CLIWrapper as a Base Class my changes related to decoding base64 data were crashing the FeatureFlags.exe process and several tests were failing.
Then I've added a extra default parameter in the GetData method so it will know if the data needs to be decoded or not.