DYN-7480 MDFiles Localization Proposal
Proposal for changes in localizing md files in Dynamo
During my research about adding the .resx files to all the projects for each language noticed that also we need to have the fallback_docs sub-folder for each language.
Then in this PR I'm proposing a change in the DynamoCore.csproj file for copying the md files to the fallback_docs inside each language folder, not sure if this change is needed due that we need to be sure that localization team has the changes for translating the md files once they are in this new location.