DYN-6897 Dynamo Button Disabled Revit
When opening Dynamo in Revit2025 and closing the SplashScreen using the Windows TaskBar the Dynamo button is disabled (there is no way to re-enable it) so you need to close Revit to be able to see the Dynamo button enabled again.
So in this fix I'm handling the case of re-enable the Dynamo button in Revit when the SplashScreen is closed using the Windows TaskBar (CloseMode.ByOther).