DYN-7480 Update Resources in DynamoCoreWpf project
I noticed that when compiling the Dynamo solution for most of the projects which contain resources in several languages (like Resources.es-ES.resx ) the ...resources.dll was generated correctly but the DynamoCoreWpf.resources.dll was missing (seems that even when the .resx files are located in \src\DynamoCoreWpf\Properties the resources are not loaded in VS2022).
Then I modified the DynamoCoreWpf.csproj adding the resources directly, in that way the DynamoCoreWpf.resources.dll was generated sucessfully in each language folder.