DYN-8275 HTML Localization Documentation Browser
I've added the html files per each language in  src\DocumentationBrowserViewExtension\Docs and were embedded as resource in the Visual Studio solution, so we can read the html based in the language selected in Dynamo.
Also I've modified the DynamoCoreWpf.csproj for copying the html files for each language from Dynamo\doc\distrib\html to Dynamo\src\DocumentationBrowserViewExtension\Docs (so when compiling the project the html will be updated if is needed).