DYN-5796-DynamoView-DisplayLocation-Bug
When launching DynamoView in a multimonitor display and closing it in the second monitor is saving the location in the DynamoSettings.xml so the next time Dynamo is opened is putting the DynamoView in previously the saved location, then if the second display is missing DynamoView is located in a position where the user cannot interact with the Window.
For fixing this problem I've added a validation for putting dynamo in the first monitor (X =0,  Y = 0) when the saved location is not valid (this fix is valid for any position horizontally and vertically).