DYN-8391 Curve Mapper Documentation
I've added the md file that contains the documentation for the Curve Mapper node.
Also I've added a resource string that contains the description of the Curve Mapper node.
Finally I've cherry-picked the commit for adding the Curve Mapper node due that was reverted some days ago (due that several tests were failing).