DYN-7480 Update Resources DynamoCoreWpf
Due that in DynamoCoreWpf.csproj we have the next tag:
<EnableDefaultEmbeddedResourceItems>False</EnableDefaultEmbeddedResourceItems>
The resources (resx) added by localization team are not loaded in VS2022 and by consequence after building the dlls are not created for the other languages.
Then in this fix I'm adding those resx files as embedded resources in DynamoCoreWpf.csproj.