DYN-8615 OpenXML Excel Exception
Probably this error is based in how the excel spreadsheet was created but basically the problem is that when the spreadsheet contains empty rows sometimes the LastChild is null so the  "Object reference not set to an instance of an object." exception is generated when trying to check any property inside the Cell instance.
For fixing this problem I just added a validation that will skip the empty row if Row.LastChild is null.