DYN-5555-Update-Sign-In Tooltiip Style
When signed in Dynamo if you press the Sign Out button was just setting the Tooltip text but the style was null, so with this fix I'm setting the Style in code behind after Sign Out button was pressed.
When Signed in no tooltip text is appearing so no need to update the content.