DYN-8801 UpdateLayout Performance
I've added the functionality of serialize into the dyn file the NodeView.Width and NodeView.Height (this values will be set on the NodeView.SizeChanged event) also when loading the graph the Width and Height will be deserialized.
Also added the functionality that we opening the graph, the Width and Height will be set to elements in the NodeView to avoid extra calls to MeasureOverride() and ArrangeOverride() this will produce some performance improvements avoiding calls to UpdateLayout().
Finally I modified the PublicAPI.Unshipped adding the new properties for Width and Height.