DYN-9414 Copy MD, HTML and XML localized files
In this fix I'm adding code in the csproj files for copying localized XML, HTML and MD files to the en-GB folder so when opening documentation for Locale = en-GB the info will be displayed in the correct language.
Also the html files copied were changed to be EmbeddedResources in this way we can load the resources in Dynamo source code.