DYN-9244 Moving Away ProgramData Fix
f you previously used Dynamo 4.0.0 and now with the previous fix try to use it, you will see that is loading packages from C:\ProgramData\Dynamo\Dynamo Core\4.0\packages.
The problem is that DynamoSettings.xml in C:\Users\tellro\AppData\Roaming\Dynamo\Dynamo Core\4.0 already contains the entry for ProgramData then you need to delete the Dynamo cache in order to verify my fix.
With this fix doesn't matter if the Dynamo cache was deleted or not now it won't be loading any package from the ProgramData path.