DYN-9244 ProgramData Samples Regressions
Due that CommonDataDirectory was set to the DynamoCore directory it was adding this folder to DirectoriesToVerifyCertificates but DynamoCore/extensions and DynamoCore/viewExtensions don't have certificates so all the built-in extensions were not loaded.
After my fix I'm not adding entires to DirectoriesToVerifyCertificates so doesn't check for extensions with certificates in DynamoCore/extensions and DynamoCore/viewExtensions then are loaded correctly.