DYN-9348-Revit-PackagePath Suggested Option Fix
For debugging purposes (when DynamoRevit is reading the info from Dynamo.config) this fix is needed otherwise when DynamoRevit calls GetDynamoPath method if the DynamoRevitVersion vs DynamoVersion doesn't match is returning null. Then with this fix DynamoPath will be always found and the Dynamo button will be shown in Revit.