DYN-9680 Crash DynamoSettingsXML
When running DynamoSandbox on C:\Dynamo was crashing due that the code was using the method Directory.GetParent() when the result was null (due that there was no two level parent), so I had to add code to make a difference between DynamoSandbox or Dynamo over a host.
Then DynamoSettings.xml will be read optionally from C:\ProgramData\Dynamo when using Sandbox and from C:\ProgramData\Autodesk\RVT 2027 when using a host like Revit 2027.