DYN-9244 Sandbox Templates Location
For DynamoSandbox was creating the templates folder in a wrong location, e.g. if the path is Dynamo\bin\AnyCPU\Debug then the templates folder was created in Dynamo\bin\AnyCPU\.
So now with this fix will be creating the templates folder in Dynamo\bin\AnyCPU\Debug, this fix will affect only DynamoSandbox. For Revit or other hosts the functionality should remain normal behavior.